+++ config/versiones.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'IEE'           => '1.15.2',
-    'BhagamApps'    => '1.15.2',
+    'IEE'           => '1.15.3',
+    'BhagamApps'    => '1.15.3',
     'User'          => '2.3.0',
-    'Inventario'    => '2.11.2',
+    'Inventario'    => '2.11.3',
     'Apps'          => '1.5.1',
     'CrudGenerator' => '1.1.0',
 ];